+++ lib/extra.dart
@@ -1,3 +1,4 @@
 export 'src/utils/utils.dart';
+export 'src/widgets/activity_barrier.dart';
 export 'src/widgets/clear_focus_background.dart';
 export 'src/widgets/faded_animated_switcher.dart';